backend/requirements.txt
fastapi
uvicorn
python-docx
pypdf
langchain
langchain-community
ollama
pydantic
python-multipart
supabase
